Cruel address impl
