+++ wallet/Cargo.toml
@@ -12,4 +12,4 @@ varint = { package = "unsigned-varint", version = "0.2.3" }
 blake2-rfc = { package = "blake2-rfc", version = "0.2.18" }
 libsecp256k1 = { version = "0.3.0" }
 derive_more = { version = "0.99.2"}
-data-encoding = { package = "data-encoding", version = "2.0.0"}
+data-encoding = { package = "data-encoding", version = "2.0.0" }
